LLM_ChatBot/requirement.txt
langchain 
python-dotenv
ipykernel 
langchain-community
pypdf
bs4 
arxiv
pymupdf
wikipedia
langchain-text-splitters 
langchain-openai
chromadb 
sentence_transformers
langchain_huggingface 
faiss-cpu
chromadb
langchain_chroma 
streamlit
langchain_groq 
langchain_core
fastapi
uvicorn 
langserve
sse_starlette 
